Projeto09-testes-unitarios-langchain-e-azure-chatgpt/requirements.txt
langchain>=0.2.11
langchain-openai>=0.1.22
langchain-core>=0.2.35
python-dotenv>=1.0.1
pytest>=8.2.0
